EES-4928 Renaming current publication subscription handlers to make it clearer
This now more closely resembles the new pattern I'll be introducing for the API subscriptions. Trying to keep these as consistent as possible without a large refactor of the current stuff
